move grid options to grid header
