+++ app/src/containers/SamplesContainer.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { useRecoilState } from "recoil";
-import { Checkbox } from "@material-ui/core";
 import styled from "styled-components";
 
 import FieldsSidebar from "../components/FieldsSidebar";
@@ -8,7 +7,6 @@ import ContainerHeader from "../components/ImageContainerHeader";
 import Samples from "../components/Samples";
 import ViewBar from "../components/ViewBar/ViewBar";
 import { scrollbarStyles } from "../components/utils";
-import { useTheme } from "../utils/hooks";
 
 import * as atoms from "../recoil/atoms";
 
@@ -27,11 +25,6 @@ const SidebarColumn = styled.div`
   overflow-x: hidden;
 `;
 
-const SidebarFooter = styled.div`
-  width: 100%;
-  padding: 0.5rem 1.5rem 0.5rem 0.5rem;
-`;
-
 const ContentColumn = styled.div`
   flex-grow: 1;
   width: 1px; // flex-related?, unset width causes the sidebar to collapse
@@ -44,35 +37,8 @@ const Container = styled.div`
   overflow: hidden;
 `;
 
-const OptionContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  background: ${({ theme }) => theme.backgroundDark};
-  box-shadow: 0 8px 15px 0 rgba(0, 0, 0, 0.43);
-  border: 1px solid #191c1f;
-  border-radius: 2px;
-  color: ${({ theme }) => theme.fontDark};
-  margin-top: 0.25rem;
-  font-weight: bold;
-  cursor: pointer;
-`;
-
-const OptionButton = styled.div`
-  display: flex;
-  justify-content: space-between;
-  background: ${({ theme }) => theme.backgroundDark};
-  box-shadow: 0 8px 15px 0 rgba(0, 0, 0, 0.43);
-  border: 1px solid #191c1f;
-  border-radius: 2px;
-  color: ${({ theme }) => theme.fontDark};
-  margin-top: 0.25rem;
-  font-weight: bold;
-`;
-
 const SamplesContainer = React.memo(() => {
   const [showSidebar, setShowSidebar] = useRecoilState(atoms.sidebarVisible);
-  const [colorByLabel, setColorByLabel] = useRecoilState(atoms.colorByLabel);
-  const theme = useTheme();
 
   return (
     <>
@@ -92,17 +58,6 @@ const SamplesContainer = React.memo(() => {
                 }}
               />
             </SidebarColumn>
-            <SidebarFooter>
-              <OptionContainer onClick={() => setColorByLabel(!colorByLabel)}>
-                <span style={{ height: "2rem", padding: "0.5rem" }}>
-                  Color by label
-                </span>
-                <Checkbox
-                  style={{ color: theme.brand }}
-                  checked={colorByLabel}
-                />
-              </OptionContainer>
-            </SidebarFooter>
           </SidebarContainer>
         ) : null}
         <ContentColumn>
